Try it now show/hide no longer uses app state (like expand collapse docs), cleaned up code we no longer need, errything fast
