+++ dynamic/react/api-app/components/endpoint.jsx
@@ -8,7 +8,7 @@ import ExpanderIcon from './expanderIcon';
 const replaceSpaces = (str) => str.replace(/\s/g, '_');
 
 // Give our endpoint an id based on its name for our clientside routing in jekyll
-const EndPointComponent = ({endpoint, apiType, id, onJumpToConsole, onConsoleVisibilityToggle, onFillConsoleSampleData, onSubmitConsoleRequest, onPostBodyInputChanged, onResetConsole, onQueryParamChanged, onPathParamChanged, onAddItemToPostbodyCollection, onRemovePostbodyCollectionItem}) => (
+const EndPointComponent = ({endpoint, apiType, id, onFillConsoleSampleData, onSubmitConsoleRequest, onPostBodyInputChanged, onResetConsole, onQueryParamChanged, onPathParamChanged, onAddItemToPostbodyCollection, onRemovePostbodyCollectionItem}) => (
     <div id={replaceSpaces(endpoint.name)}>
         <div className={'endpoint-summary'}>
             <h2>{endpoint.name}</h2>
@@ -20,7 +20,6 @@ const EndPointComponent = ({endpoint, apiType, id, onJumpToConsole, onConsoleVis
                         onClick={
                             () => {
                                 $(`#${replaceSpaces(endpoint.name)}-console-body`).collapse('show');
-                                onJumpToConsole(id);
                             }
                         }
                     >{'Try it now!'}</a>
@@ -49,10 +48,10 @@ const EndPointComponent = ({endpoint, apiType, id, onJumpToConsole, onConsoleVis
             <div>
                 <div className={'try-it-now-header'} data-target={`#${replaceSpaces(endpoint.name)}-console-body`} data-toggle={'collapse'} id={replaceSpaces(`${endpoint.name}-console`)} onClick={
                     () => {
-                        onConsoleVisibilityToggle(id);
+                        $(`#${replaceSpaces(endpoint.name)}-console-icon`).toggleClass('rotate');
                     }
                 }>
-                    <div className={'documentation-expand-icon' + (endpoint.apiConsoleVisible ? ' rotate' : '')} style={{display: 'inline-block', marginRight: '5px'}}>
+                    <div className={'documentation-expand-icon'} id={`${replaceSpaces(endpoint.name)}-console-icon`} style={{display: 'inline-block', marginRight: '5px'}}>
                         <ExpanderIcon startPosition={'DOWN'}/>
                     </div>
                     <span>{'Try it now! '}</span>
@@ -101,14 +100,11 @@ EndPointComponent.propTypes = {
             body: React.PropTypes.oneOfType([
                 React.PropTypes.object, React.PropTypes.array
             ]).isRequired
-        }),
-        apiConsoleVisible: React.PropTypes.bool.isRequired
+        })
     }).isRequired,
     id: React.PropTypes.number.isRequired,
     onAddItemToPostbodyCollection: React.PropTypes.func.isRequired,
-    onConsoleVisibilityToggle: React.PropTypes.func.isRequired,
     onFillConsoleSampleData: React.PropTypes.func.isRequired,
-    onJumpToConsole: React.PropTypes.func.isRequired,
     onPathParamChanged: React.PropTypes.func.isRequired,
     onPostBodyInputChanged: React.PropTypes.func.isRequired,
     onQueryParamChanged: React.PropTypes.func.isRequired,
