+++ dynamic/react/get-started/reducers/reducer.js
@@ -13,10 +13,7 @@ export default (state, action) => {
     case actionTypes.REMOVE_ITEM_FROM_POST_BODY_COLLECTION:
     case actionTypes.FILL_REQUEST_SAMPLE_DATA:
     case actionTypes.AUTH_KEY_CHANGED:
-    case actionTypes.CONSOLE_VISIBILITY_TOGGLED:
-    case actionTypes.JUMP_TO_CONSOLE:
     case actionTypes.RESET_CONSOLE:
-    case actionTypes.TOGGLE_DOCUMENTATION_ITEM_VISIBILITY:
         const newEndpointState = apiConsoleReducer(state[action.endpointId], action);
 
         return state.map((ep, i) => {
