+++ dynamic/react/get-started/swaggerToGetStartedState.js
@@ -4,9 +4,7 @@ import path from 'path';
 
 export default (filename, callback) => {
     new SwaggerParser().dereference(path.join(__dirname, '..', '..', 'swagger', filename)).then((swaggerDoc) => {
-        const getStartedState = parseSwaggerUi(swaggerDoc, null).apiEndpoints.map((endpoint) => {
-            return {...endpoint, apiConsoleVisible: true};
-        });
+        const getStartedState = parseSwaggerUi(swaggerDoc, null).apiEndpoints;
 
         callback(getStartedState);
     });
