+++ dynamic/react/shared/actionTypes.js
@@ -8,8 +8,5 @@ export default {
     REMOVE_ITEM_FROM_POST_BODY_COLLECTION: 'REMOVE_ITEM_FROM_POST_BODY_COLLECTION',
     FILL_REQUEST_SAMPLE_DATA: 'FILL_REQUEST_SAMPLE_DATA',
     AUTH_KEY_CHANGED: 'AUTH_KEY_CHANGED',
-    CONSOLE_VISIBILITY_TOGGLED: 'CONSOLE_VISIBILITY_TOGGLED',
-    JUMP_TO_CONSOLE: 'JUMP_TO_CONSOLE',
-    RESET_CONSOLE: 'RESET_CONSOLE',
-    TOGGLE_DOCUMENTATION_ITEM_VISIBILITY: 'TOGGLE_DOCUMENTATION_ITEM_VISIBILITY'
+    RESET_CONSOLE: 'RESET_CONSOLE'
 };
